add user login and register

+++ resource/src/main/java/com/cpu/news/controller/NewsController.java
@@ -4,6 +4,8 @@ package com.cpu.news.controller;
 import com.cpu.department.DepartmentService;
 import com.cpu.news.News;
 import com.cpu.news.NewsService;
+import com.cpu.user.User;
+import com.cpu.utils.SignInUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -42,13 +44,20 @@ public class NewsController {
 
     @RequestMapping("/")
     public String newsList(Model model) {
-        model.addAttribute("newsList", service.findAll());
+        User user = SignInUtils.getInstance().getCurrentUser();
+        if (user == null) {
+            return "redirect:/";
+        }
+        model.addAttribute("newsList",user.getIdNumber().equals("admin")? service.findAll():service.findByUserId(user.getId()));
         return "news/news_list";
     }
 
 
     @RequestMapping("/create")
     public String create(Model model) {
+        if (SignInUtils.getInstance().getCurrentUser() == null) {
+            return "redirect:/";
+        }
         News news = new News();
         model.addAttribute("departments", departmentService.findAll());
         model.addAttribute("news", news);
@@ -57,7 +66,9 @@ public class NewsController {
 
     @RequestMapping("/view")
     public String view(@RequestParam long id, Model model) {
-
+        if (SignInUtils.getInstance().getCurrentUser() == null) {
+            return "redirect:/";
+        }
         News news = service.findOne(id);
         model.addAttribute("news", news);
         return "news/news_view";
@@ -66,6 +77,9 @@ public class NewsController {
 
     @RequestMapping("/edit")
     public String edit(@RequestParam long id, Model model) {
+        if (SignInUtils.getInstance().getCurrentUser() == null) {
+            return "redirect:/";
+        }
         News news = service.findOne(id);
         model.addAttribute("news", news);
         model.addAttribute("departments", departmentService.findAll());
@@ -74,6 +88,9 @@ public class NewsController {
 
     @RequestMapping("/delete")
     public String delete(@RequestParam long id, RedirectAttributes redir) {
+        if (SignInUtils.getInstance().getCurrentUser() == null) {
+            return "redirect:/";
+        }
         service.delete(id);
         return "redirect:/page/news/";
     }
@@ -81,9 +98,11 @@ public class NewsController {
     @PostMapping("/save")
     public String saveNews(@ModelAttribute News news, @RequestParam MultipartFile[] files, RedirectAttributes redir) {
         try {
-            News savedNews = service.findOne(news.getId());
-            news.setImage(savedNews.getImage());
-            news.setFilePath(savedNews.getFilePath());
+            if(news.getId()!=null){
+                News savedNews = service.findOne(news.getId());
+                news.setImage(savedNews.getImage());
+                news.setFilePath(savedNews.getFilePath());
+            }
             for (MultipartFile file : files) {
                 if (!file.isEmpty()) {
                     continue;
@@ -103,9 +122,10 @@ public class NewsController {
                     byte[] imageBytes = Base64.getEncoder().encode(file.getBytes());
                     news.setImage(new String(imageBytes));
                 }
-                news.setDepartment(departmentService.findOne(news.getDeptId()));
-                service.save(news);
             }
+            news.setDepartment(departmentService.findOne(news.getDeptId()));
+            news.setUserId(SignInUtils.getInstance().getCurrentUser().getId());
+            service.save(news);
 
         } catch (IOException e) {
             e.printStackTrace();
